Moving extensions from EndpointRouteBuilder
